fb: Post: Introduce `$take_content` argument in `getTimelinePosts()`
Allow the user to get not only the posts list, but also the content of
each post directly within a single API call. This one better be
accompanied with a $limit argument to efficiently get the latest post
only, which is what our users want. Let's add the $limit argument in the
next patch.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>