fb: Post: Replace old cache mechanism in `getTimelineYears()`
`getTimelineYears()` always fetches the endpoint online, then it sets
the cache based on the fetch result. On the other hand, the
`getTimelinePosts()` method always tries to read the cache that
`getTimelineYears()` sets.

Now, simplify the caching mechanism and make the `getTimelineYears()`
cache private to itself. This also means that the endpoint
"action=getTimelineYears" will utilize the cache.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20230509104658.70953-3-ammarfaizi2@gnuweeb.org
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>