fb: Post: Fix the wrong content type detection (photo issue)
Michael reports that using the recent API, photo contents in the
HonkaiStarRail.ID timeline are detected as a "text".

Using the current parser, it detects those contents as a text with a
link preview (where the preview is the post to the photo). Just visit
the photo link if there exists before continue trying to parse it as a
photo. Also, make sure we try to parse as a photo first before text.

Cc: Michael William Jonathan <moe@gnuweeb.org>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>