fb: Introduce orig_base_url property
A preparation patch to support high quality image scrapper. To scrape
the high quality image, it's required to visit the Facebook Desktop
version because the mbasic version currently doesn't have it.

Allow the core trait to temporary change the base URL to visit the
Facebook Desktop version and revert the base URL back to mbasic when it
already completes fetching the HD image URL.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>